[FIX] E-invoice-in fix bug
