+++ l10n_it_ade/static/description/index.html
@@ -16,7 +16,7 @@ It defines the structures by Italian IRS (Tax Authority) to manage
 all fiscal communications.
 Inside there are xml schema files used by FatturaPA, EInvoice and VAT settlement.
 </p><p align="justify">
-This module requires `PyXB 1.2.4 &lt;http://pyxb.sourceforge.net/&gt;`__ or `PyXB 1.2.5 &lt;http://pyxb.sourceforge.net/&gt;`__
+This module requires <a href="http://pyxb.sourceforge.net/">PyXB 1.2.4</a> or <a href="http://pyxb.sourceforge.net/">PyXB 1.2.5</a>
 </p></p>
         </div>
         <div class="oe_span6">
@@ -34,7 +34,7 @@ Fattura Elettronica B2B, Liquidazione IVA elettronica e Comunicazione IVA.
 </p><p align="justify">
 Tutti i moduli della localizzazione italiana che generano file xml dipendenti
 dallo schema dell'Agenzia delle Entrate devono dichiare il modulo
-`l10n_it_ade &lt;https://github.com/zeroincombenze/l10n-italy/tree/10.0/l10n_it_ade&gt;`__ come dipendenza.
+<a href="https://github.com/zeroincombenze/l10n-italy/tree/10.0/l10n_it_ade/">l10n_it_ade</a> come dipendenza.
 </p><p align="justify">
 Per maggiori informazioni contattare
 l'ultimo autore: Antonio M. Vigliotti &lt;antoniomaria.vigliotti@gmail.com&gt;.
