+++ l10n_it_einvoice_out/static/description/index.html
@@ -40,7 +40,7 @@ Il modulo &#232; destinato a tutte le aziende che dal 2019 dovranno emettere fat
     Normativa:
 </code>
 </p><p align="justify">
-Le leggi inerenti la fattura elettronica sono numerose. Potete consultare la `normativa fattura elettronica &lt;https://www.fatturapa.gov.it/export/fatturazione/it/normativa/norme.htm&gt;`__
+Le leggi inerenti la fattura elettronica sono numerose. Potete consultare la <a href="https://www.fatturapa.gov.it/export/fatturazione/it/normativa/norme.htm/">normativa fattura elettronica</a>
 </p><p align="justify">
 Per maggiori info leggere le informazioni relative al modulo l10n_it_einvoice_base
 </p></p>
