+++ l10n_it_fiscal_payment_term/static/description/index.html
@@ -11,7 +11,7 @@
             <p class="oe_mt16" align="justify"><h1>Italian Fiscal Payment Terms</h1>
 <p align="justify">
 E-Invoice and FatturaPA Payment Terms as per
-`Italian Government Website &lt;http://fatturapa.gov.it&gt;`__
+<a href="http://fatturapa.gov.it/">Italian Government Website</a>
 </p></p>
         </div>
         <div class="oe_span6">
@@ -21,7 +21,7 @@ E-Invoice and FatturaPA Payment Terms as per
             <p class="oe_mt16" align="justify"><h1>Termini di Pagamento Fiscali </h1>
 <p align="justify">
 Termini di pagamento fiscali per Fattura Elettronica e FatturaPA come da 
-`Disposizioni governative FatturaPA &lt;http://fatturapa.gov.it&gt;`__
+<a href="http://fatturapa.gov.it/">Disposizioni governative FatturaPA</a>
 </p></p>
         </div>
     </div>
